FEM-realtime: socket io simple conn
