+++ 03-frontend-masters/11-intro-to-realtime/websockets/exercise-socketio/backend/server.js
@@ -19,11 +19,30 @@ const server = http.createServer((request, response) => {
   });
 });
 
-/*
- *
- * Code goes here
- *
- */
+//// IMPL
+
+const io = new Server(server, {}); // reuse server
+io.on("connection", (socket) => {
+  console.log(`connected: ${socket.id}`); // unique per socket
+
+  socket.emit("msg:get", { msg: getMsgs() });
+
+  socket.on("msg:post", (data) => {
+    msg.push({
+      user: data.user,
+      text: data.text,
+      time: Date.now(),
+    });
+
+    io.emit("msg:get", { msg: getMsgs() });
+  });
+
+  socket.on("disconnect", () => {
+    console.log(`disconnected: ${socket.id}`);
+  });
+});
+
+////
 
 const port = process.env.PORT || 8080;
 server.listen(port, () =>
