+++ 03-frontend-masters/11-intro-to-realtime/websockets/exercise-socketio/frontend/socketio-chat.js
@@ -1,15 +1,29 @@
-// a global called "io" is being loaded separately
+// a global called "io" is being loaded separately (in index.html)
 
 const chat = document.getElementById("chat");
 const msgs = document.getElementById("msgs");
 const presence = document.getElementById("presence-indicator");
 let allChat = [];
 
-/*
- *
- * Code goes here
- *
- */
+// a
+
+const socket = io("http://localhost:8080");
+
+socket.on("connect", () => {
+  console.log("Connected");
+  presence.innerText = "🎾";
+});
+
+socket.on("disconnect", () => {
+  presence.innerText = "🏓";
+});
+
+socket.on("msg:get", (data) => {
+  allChat = data.msg;
+  render();
+});
+
+//
 
 chat.addEventListener("submit", function (e) {
   e.preventDefault();
@@ -18,11 +32,9 @@ chat.addEventListener("submit", function (e) {
 });
 
 async function postNewMsg(user, text) {
-  /*
-   *
-   * Code goes here
-   *
-   */
+  const data = { user, text };
+
+  socket.emit("msg:post", data);
 }
 
 function render() {
