--- 03-frontend-masters/11-intro-to-realtime/websockets/socketio/frontend/style.css
@@ -1,19 +1,0 @@
-.page-column {
-  max-width: 600px;
-  margin: 30px auto;
-}
-
-h1 {
-  text-align: center;
-  color: #26a69a;
-}
-
-.btn-box {
-  text-align: center;
-}
-
-#presence-indicator {
-  position: absolute;
-  top: 5px;
-  right: 15px;
-}
